fix(web): split the board-dispatch sentinels out of the server-only tasks module
`RunWithAgentMenu` is a `'use client'` component and imports
`RUN_AGENT_AMBIGUOUS` / `RUN_ALREADY_IN_FLIGHT` / `RUN_NO_AGENT` as
VALUES from `@/lib/api/tasks`, which starts with `import 'server-only'`.
That pulls the server-only guard into the client bundle and fails
`next build`:

    x You're importing a module that depends on "server-only".
      ./src/lib/api/tasks.ts
      ./src/components/tasks/RunWithAgentMenu.tsx
      ./src/components/tasks/TaskDetailClient.tsx

`tsc --noEmit` is clean, so nothing before the production build catches
it. develop HEAD (e64d3a12) is red on exactly this: the k8s-build web
image cannot compile, so develop cannot deploy, and the e2e prod-build
job would die the same way.

Fix is the house `.shared.ts` split already used by goals/agents/
billing/credits: the sentinels and `RunCandidateAgent` move to
`tasks.shared.ts` (no server dependency), `tasks.ts` re-exports them so
server-side callers keep one import site, and the client component
imports from the shared module.

Verified with a real `turbo build --filter=ever-works-web` (passes) —
not just tsc.
